requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ML/DL Dependencies
torch>=2.4.0
torchvision>=0.19.0
ultralytics>=8.3.0

# Computer Vision
opencv-python>=4.10.0
numpy>=2.0.0

# DeepSORT Tracking
deep-sort-realtime==1.3.2

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Async Support
aiofiles==23.2.1

# Database
tortoise-orm[asyncpg]==0.21.7
aerich==0.7.2

# Utilities
python-dotenv==1.0.1
setuptools>=75.0.0
